+++ app/shortcutsdialog.cpp
@@ -1,0 +1,27 @@
+#include "shortcutsdialog.h"
+#include "ui_shortcutsdialog.h"
+
+ShortcutsDialog::ShortcutsDialog(const Shortcuts *shortcuts, QWidget *parent) : QDialog(parent), ui(new Ui::ShortcutsDialog) {
+  ui->setupUi(this);
+
+  auto data = shortcuts->describe();
+  ui->tableWidget->setRowCount(data.size());
+  ui->tableWidget->setColumnCount(2);
+  ui->tableWidget->verticalHeader()->hide();
+  ui->tableWidget->horizontalHeader()->hide();
+  ui->tableWidget->horizontalHeader()->setSectionResizeMode(0, QHeaderView::Stretch);
+  ui->tableWidget->horizontalHeader()->setSectionResizeMode(1, QHeaderView::Stretch);
+
+  for (int row = 0; row < data.size(); row++) {
+    auto item1 = new QTableWidgetItem(data[row].first);
+    auto item2 = new QTableWidgetItem(data[row].second);
+    item1->setFlags(Qt::ItemIsEnabled);
+    item2->setFlags(Qt::ItemIsEnabled);
+    ui->tableWidget->setItem(row, 0, item1);
+    ui->tableWidget->setItem(row, 1, item2);
+  }
+}
+
+ShortcutsDialog::~ShortcutsDialog() {
+  delete ui;
+}